- Added data/gambling.yml configuration - gambling config contains everything botconfig had related to currency/gambling and much more - your previous configs will be migrated to .yml file on first start - config changes via .config command (or changing the file and doing configreload) are immediately applied without need for restart - added .config command which will let you   - list configs   - list all config properties   - get config property value and description/comment   - set config property value - added .configreload command which lets you reload specified config - currency decay functionality is improved   - you can specify minimum amount of currency user must have in order to experience decay   - you can specify maximum amount of decay that can be deduced from user's currency   - you can specify an interval in hours at which the decay runs - Several multipliers and settings added for .waifu related commands - You can specify custom waifu items from the .yml file - Slight changes to required numbers of affinity changes required for certain titles - Yaml serializer will no longer save emojis as codepoints - .bce will be completely removed in a future patch and replaced with .config related commands - Removed obsolete botsettings - SettingsServiceBase<TSetting> now implements ISettings - Fixed setting nested config property - renamed WizBotTopLevelModule to WizBotModule - added a test for codepoint -> emoji conversion - Updated a few comments on bot config - cleanup of some old comments/unused code

+++ WizBot.Core/Services/Settings/BotSettingsMigrator.cs
@@ -43,7 +43,7 @@ namespace WizBot.Core.Services
 
             using var checkMigratedCommand = conn.CreateCommand();
             checkMigratedCommand.CommandText = "UPDATE BotConfig SET HasMigratedBotSettings = 1 WHERE HasMigratedBotSettings = 0;";
-            var changedRows = (int)checkMigratedCommand.ExecuteNonQuery();
+            var changedRows = checkMigratedCommand.ExecuteNonQuery();
             if (changedRows == 0)
                 return;
